refactor | c&i : split into modules
